Enhance particle-material interaction for hard drops and line clears
- Extracted particle interaction WGSL logic to `particleMaterialInteraction.ts`.
- Integrated particle interactions into `premiumBlocks.ts` and `pbrBlocks.ts` shaders.
- Added a `particleHitTimer` to `effects.ts` with exponential decay to control interaction intensity dynamically.
- Triggers particle hit interactions during `onHardDrop` and `onLineClear` in `viewGameEvents.ts`.
- Maps the dynamic intensity value into the fragment uniform buffer at byte offset 84 (`particleIntensity`).
- Calculates the correct `materialType` for PBR blocks dynamically from `metalMask` and `glassMask` to ensure accurate shader reactions (glass refraction, gold/chrome specular flashes).
