+++ neon_bricklayers_journal.md
@@ -45,3 +45,4 @@
 - "Wired in the gold cracked-glass texture as a subtle dynamic overlay during the hard drop shockwave post-processing effect, extracting bright highlights from the block texture to simulate a shattering impact."
 - "Implemented explicit Fresnel Rim Lighting on the gold glass blocks! Added a `fresnelParams` uniform in `viewWebGPU.ts` and updated the PBR fragment shader (`pbrBlocks.ts`) to compute a Fresnel term based on the view direction. Wired a `setFresnelBoost()` hook in `game.ts` to dramatically intensify the gold edges when a Hard Drop occurs, creating an awesome additive rim flare during impacts."
 - "Implemented a 'Neon Burst' Radial Distortion & Glow effect on Hard Drops, giving the screen a crunchy cyan/magenta hit and satisfying visual decay."
+- "Implemented Particle-Material Interaction! Extracted interaction logic into a shared WGSL module and injected it into `pbrBlocks.ts` and `premiumBlocks.ts`. Added a decaying `particleHitTimer` to `VisualEffects` that triggers on hard drops and line clears, passing it as `particleIntensity` to the shaders. Now, when particles hit the blocks, they react with spectacular, material-aware visual responses like refraction on glass, specular flashes on gold/chrome, and neon bursts on cyber blocks."
